+++ src/widgets/terminal/node.rs
@@ -11,7 +11,8 @@ use crate::style::{
 use crate::widgets::ScrollEvent;
 
 use super::events::{
-    MouseModeState, TerminalInputEvent, TerminalKeyModes, TerminalSelection, TerminalSelectionEvent,
+    MouseModeState, TerminalInputEvent, TerminalKeyModes, TerminalPasteShortcutBehavior,
+    TerminalSelection, TerminalSelectionEvent,
 };
 use super::layout::terminal_content_layout;
 
@@ -28,6 +29,7 @@ pub(crate) struct TerminalNode {
     pub selection_style: StyleSlot,
     pub mouse_mode: MouseModeState,
     pub key_modes: TerminalKeyModes,
+    pub paste_shortcut_behavior: TerminalPasteShortcutBehavior,
     pub on_selection: Option<Callback<TerminalSelectionEvent>>,
     pub on_mouse_forward: Option<Callback<Vec<u8>>>,
     pub style: Style,